Solution of Physics-based Bayesian Inverse Problems with Deep Generative Priors
Abstract
Inverse problems are notoriously difficult to solve because they can have no solutions,
multiple solutions, or have solutions that vary significantly in response to small perturbations
in measurements. Bayesian inference, which poses an inverse problem as a stochastic inference
problem, addresses these difficulties and provides quantitative estimates of the inferred field
and the associated uncertainty. However, it is difficult to employ when inferring vectors of
large dimensions, and/or when prior information is available through previously acquired
samples. In this paper, we describe how deep generative adversarial networks can be used to
represent the prior distribution in Bayesian inference and overcome these challenges. We
apply these ideas to inverse problems that are diverse in terms of the governing physical
principles, sources of prior knowledge, type of measurement, and the extent of available
information about measurement noise. In each case we apply the proposed approach to infer
the most likely solution and quantitative estimates of uncertainty.