Plot2Spectra: an Automatic Spectra Extraction Tool
Abstract
Different types of spectroscopies, such as X-ray absorption near edge structure (XANES)
and Raman spectroscopy, play a very important role in analyzing the characteristics of
different materials. In scientific literature, XANES/Raman data are usually plotted in line
graphs which is a visually appropriate way to represent the information when the end-
user is a human reader. However, such graphs are not conducive to direct programmatic
analysis due to the lack of automatic tools. In this paper, we develop a plot digitizer,
named Plot2Spectra, to extract data points from spectroscopy graph images in an au-
tomatic fashion, which makes it possible for large scale data acquisition and analysis.
Specifically, the plot digitizer is a two-stage framework. In the first axis alignment stage,
we adopt an anchor-free detector to detect the plot region and then refine the detected
bounding boxes with an edge-based constraint to locate the position of two axes. We
also apply scene text detector to extract and interpret all tick information below the x-
axis. In the second plot data extraction stage, we first employ semantic segmentation to
separate pixels belonging to plot lines from the background, and from there, incorporate
optical flow constraints to the plot line pixels to assign them to the appropriate line (data
instance) they encode. Extensive experiments are conducted to validate the effectiveness
of the proposed plot digitizer, which shows that such a tool could help accelerate the
discovery and machine learning of materials properties.